+++ src/bl64-cnt-env.bash
@@ -2,7 +2,7 @@
 # BashLib64 / Module / Globals / Interact with container engines
 #######################################
 
-export BL64_CNT_VERSION='1.9.1'
+export BL64_CNT_VERSION='1.9.2'
 
 # Optional module. Not enabled by default
 export BL64_CNT_MODULE="$BL64_VAR_OFF"