+++ src/bl64-pkg-env.bash
@@ -2,7 +2,7 @@
 # BashLib64 / Module / Globals / Manage native OS packages
 #######################################
 
-export BL64_PKG_VERSION='4.0.0'
+export BL64_PKG_VERSION='4.0.1'
 
 export BL64_PKG_MODULE="$BL64_VAR_OFF"