+++ src/bl64-rbac-env.bash
@@ -2,7 +2,7 @@
 # BashLib64 / Module / Globals / Manage role based access service
 #######################################
 
-export BL64_RBAC_VERSION='1.12.0'
+export BL64_RBAC_VERSION='1.12.1'
 
 export BL64_RBAC_MODULE="$BL64_VAR_OFF"