+++ src/bl64-lib-env.bash
@@ -2,7 +2,7 @@
 # BashLib64 / Module / Globals / Setup script run-time environment
 #######################################
 
-export BL64_VERSION='12.4.0'
+export BL64_VERSION='12.5.0'
 
 # Declare imported variables
 export LANG